requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
gspread>=5.12.0
gspread-formatting>=1.1.2
opencv-python>=4.8.0
albumentations>=1.3.1
numpy>=1.24.0
